Tweaking and formatting
